Refactor files
